delegate playwright code to different files
